almostFinished

+++ Shoppp/ViewModels/ArticlePurchaseHistoryViewModel.cs
@@ -4,6 +4,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
+using OnlineShopping.HelperUser;
+
 namespace OnlineShopping.ViewModels
 {
     public class ArticlePurchaseHistoryViewModel
@@ -12,7 +14,7 @@ namespace OnlineShopping.ViewModels
         public Artikl Artikl { get; set; }
         public DateTime PurchaseDate { get; set; }
         public int Amount { get; set; }
-        public double TotalPrice => Artikl.CijenaArtikla * Amount;
+        public decimal TotalPrice => Artikl.CijenaArtikla * Amount;
     }
    
 }
